+++ frontend/src/utils/FieldValidationRules.ts
